Refaktoring parserů bankovních e-mailových avíz
  - přidává společný AbstractBankEmailNoticeParser
  - přesouvá sdílené helpery pro normalizaci, regexy, datumy, částky, účty a symboly
  - zachovává bank-specific supports() a parse() logiku v konkrétních parserech
  - sjednocuje regex parser se stejnou pomocnou vrstvou
